Clean up tests in legend-pure-m2-dsl-diagram-grammar
