+++ legend-pure-dsl/legend-pure-dsl-diagram/legend-pure-m2-dsl-diagram-grammar/pom.xml
@@ -109,11 +109,5 @@
             <artifactId>junit</artifactId>
             <scope>test</scope>
         </dependency>
-        <dependency>
-            <groupId>org.eclipse.collections</groupId>
-            <artifactId>eclipse-collections-testutils</artifactId>
-            <scope>test</scope>
-        </dependency>
     </dependencies>
-
 </project>
