+++ legend-pure-dsl/legend-pure-dsl-diagram/legend-pure-m2-dsl-diagram-grammar/src/test/java/org/finos/legend/pure/m2/dsl/diagram/TestDiagramParsing.java
@@ -12,19 +12,17 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package org.finos.legend.pure.m2.dsl.diagram.test.incremental;
+package org.finos.legend.pure.m2.dsl.diagram;
 
 import org.eclipse.collections.api.RichIterable;
-import org.eclipse.collections.api.block.predicate.Predicate;
 import org.finos.legend.pure.m3.coreinstance.meta.pure.diagram.AssociationView;
 import org.finos.legend.pure.m3.coreinstance.meta.pure.diagram.TypeView;
 import org.finos.legend.pure.m3.coreinstance.meta.pure.metamodel.ReferenceUsage;
 import org.finos.legend.pure.m3.coreinstance.meta.pure.metamodel.relationship.Association;
 import org.finos.legend.pure.m3.coreinstance.meta.pure.metamodel.type.Class;
 import org.finos.legend.pure.m3.navigation.Instance;
+import org.finos.legend.pure.m3.navigation.M3Paths;
 import org.finos.legend.pure.m3.navigation.M3Properties;
-import org.finos.legend.pure.m3.serialization.grammar.Parser;
-import org.finos.legend.pure.m3.serialization.grammar.ParserLibrary;
 import org.finos.legend.pure.m3.tests.AbstractPureTestWithCoreCompiled;
 import org.finos.legend.pure.m4.coreinstance.CoreInstance;
 import org.finos.legend.pure.m4.coreinstance.SourceInformation;
@@ -33,8 +31,6 @@ import org.junit.After;
 import org.junit.Assert;
 import org.junit.BeforeClass;
 import org.junit.Test;
-
-import java.lang.reflect.Field;
 
 public class TestDiagramParsing extends AbstractPureTestWithCoreCompiled
 {
@@ -51,259 +47,155 @@ public class TestDiagramParsing extends AbstractPureTestWithCoreCompiled
         runtime.delete("testModel.pure");
     }
 
-    private Parser getRuntimeDiagramParser() throws NoSuchFieldException, IllegalAccessException
-    {
-        Field field = runtime.getSourceRegistry().getClass().getDeclaredField("parserLibrary");
-        field.setAccessible(true);
-        ParserLibrary parserLibrary = (ParserLibrary) field.get(runtime.getSourceRegistry());
-        Parser parser = parserLibrary.getParser("Diagram");
-
-        return parser;
-    }
-
     @Test
-    public void testDiagramWithInvalidGeometry() throws NoSuchFieldException, IllegalAccessException
+    public void testDiagramWithInvalidGeometry()
     {
-        Parser parser = getRuntimeDiagramParser();
         // Empty geometry
-        try
-        {
-            compileTestSource("###Diagram\nDiagram test::pure::TestDiagram() {}");
-            Assert.fail("Expected a parser error");
-        }
-        catch (Exception e)
-        {
-            assertPureException(PureParserException.class, "expected: one of {WIDTH, HEIGHT} found: ')'", 2, 33, e);
-        }
+        PureParserException e1 = Assert.assertThrows(PureParserException.class, () -> compileTestSource("###Diagram\nDiagram test::pure::TestDiagram() {}"));
+        assertPureException(PureParserException.class, "expected: one of {WIDTH, HEIGHT} found: ')'", 2, 33, e1);
 
         // Width but no height
-        try
-        {
-            compileTestSource("###Diagram\nDiagram test::pure::TestDiagram(width=10.0) {}");
-            Assert.fail("Expected a parser error");
-        }
-        catch (Exception e)
-        {
-            assertPureException(PureParserException.class, "expected: ',' found: ')'", 2, 43, e);
-        }
+        PureParserException e2 = Assert.assertThrows(PureParserException.class, () -> compileTestSource("###Diagram\nDiagram test::pure::TestDiagram(width=10.0) {}"));
+        assertPureException(PureParserException.class, "expected: ',' found: ')'", 2, 43, e2);
 
         // Height but no width
-        try
-        {
-            compileTestSource("###Diagram\nDiagram test::pure::TestDiagram(height=10.0) {}");
-            Assert.fail("Expected a parser error");
-        }
-        catch (Exception e)
-        {
-            assertPureException(PureParserException.class, "expected: ',' found: ')'", 2, 44, e);
-        }
+        PureParserException e3 = Assert.assertThrows(PureParserException.class, () -> compileTestSource("###Diagram\nDiagram test::pure::TestDiagram(height=10.0) {}"));
+        assertPureException(PureParserException.class, "expected: ',' found: ')'", 2, 44, e3);
 
         // Wrong property first
-        try
-        {
-            compileTestSource("###Diagram\nDiagram test::pure::TestDiagram(junk=10.0) {}");
-            Assert.fail("Expected a parser error");
-        }
-        catch (Exception e)
-        {
-            assertPureException(PureParserException.class, "token recognition error at: '='", 2, 37, e);
-        }
+        PureParserException e4 = Assert.assertThrows(PureParserException.class, () -> compileTestSource("###Diagram\nDiagram test::pure::TestDiagram(junk=10.0) {}"));
+        assertPureException(PureParserException.class, "token recognition error at: '='", 2, 37, e4);
 
         // Wrong property second
-        try
-        {
-            compileTestSource("###Diagram\nDiagram test::pure::TestDiagram(width=10.0, junk=10.0) {}");
-            Assert.fail("Expected a parser error");
-        }
-        catch (Exception e)
-        {
-            assertPureException(PureParserException.class, "expected: HEIGHT found: 'junk'", 2, 45, e);
-        }
+        PureParserException e5 = Assert.assertThrows(PureParserException.class, () -> compileTestSource("###Diagram\nDiagram test::pure::TestDiagram(width=10.0, junk=10.0) {}"));
+        assertPureException(PureParserException.class, "expected: HEIGHT found: 'junk'", 2, 45, e5);
 
         // Extra property
-        try
-        {
-            compileTestSource("###Diagram\nDiagram test::pure::TestDiagram(width=10.0, height=10.0, junk=13.2) {}");
-            Assert.fail("Expected a parser error");
-        }
-        catch (Exception e)
-        {
-            assertPureException(PureParserException.class, "expected: ')' found: ','", 2, 56, e);
-        }
+        PureParserException e6 = Assert.assertThrows(PureParserException.class, () -> compileTestSource("###Diagram\nDiagram test::pure::TestDiagram(width=10.0, height=10.0, junk=13.2) {}"));
+        assertPureException(PureParserException.class, "expected: ')' found: ','", 2, 56, e6);
     }
 
     @Test
-    public void testDiagramWithInvalidTypeView() throws NoSuchFieldException, IllegalAccessException
+    public void testDiagramWithInvalidTypeView()
     {
-        Parser parser = getRuntimeDiagramParser();
         // Empty type view
-        try
-        {
-            compileTestSource("###Diagram\n" +
-                    "Diagram test::pure::TestDiagram(width=10.0, height=10.0)\n" +
-                    "{\n" +
-                    "    TypeView TestClass_1()\n" +
-                    "}\n");
-            Assert.fail("Expected a parser error");
-        }
-        catch (Exception e)
-        {
-            assertPureException(PureParserException.class, "expected: one of {TYPE, STEREOTYPES_VISIBLE, ATTRIBUTES_VISIBLE, ATTRIBUTE_STEREOTYPES_VISIBLE, ATTRIBUTE_TYPES_VISIBLE, COLOR, LINE_WIDTH, POSITION, WIDTH, HEIGHT} found: ')'", 4, 26, e);
-        }
+        PureParserException e1 = Assert.assertThrows(PureParserException.class, () -> compileTestSource(
+                "###Diagram\n" +
+                        "Diagram test::pure::TestDiagram(width=10.0, height=10.0)\n" +
+                        "{\n" +
+                        "    TypeView TestClass_1()\n" +
+                        "}\n"));
+        assertPureException(PureParserException.class, "expected: one of {TYPE, STEREOTYPES_VISIBLE, ATTRIBUTES_VISIBLE, ATTRIBUTE_STEREOTYPES_VISIBLE, ATTRIBUTE_TYPES_VISIBLE, COLOR, LINE_WIDTH, POSITION, WIDTH, HEIGHT} found: ')'", 4, 26, e1);
 
         // Type view with one bogus property
-        try
-        {
-            compileTestSource("###Diagram\n" +
-                    "Diagram test::pure::TestDiagram(width=10.0, height=10.0)\n" +
-                    "{\n" +
-                    "    TypeView TestClass_1(junk=13.6)\n" +
-                    "}\n");
-            Assert.fail("Expected a parser error");
-        }
-        catch (Exception e)
-        {
-            assertPureException(PureParserException.class, "token recognition error at: '='", 4, 30, e);
-        }
+        PureParserException e2 = Assert.assertThrows(PureParserException.class, () -> compileTestSource(
+                "###Diagram\n" +
+                        "Diagram test::pure::TestDiagram(width=10.0, height=10.0)\n" +
+                        "{\n" +
+                        "    TypeView TestClass_1(junk=13.6)\n" +
+                        "}\n"));
+        assertPureException(PureParserException.class, "token recognition error at: '='", 4, 30, e2);
 
         // Type view with all valid properties but one
-        try
-        {
-            compileTestSource("###Diagram\n" +
-                    "Diagram test::pure::TestDiagram(width=10.0, height=10.0)\n" +
-                    "{\n" +
-                    "    TypeView TestClass_1(stereotypesVisible=true, attributesVisible=true,\n" +
-                    "                         attributeStereotypesVisible=true, attributeTypesVisible=true,\n" +
-                    "                         color=#FFFFCC, lineWidth=1.0,\n" +
-                    "                         position=(874.0, 199.46875), width=353.0, height=57.1875)\n" +
-                    "}\n");
-            Assert.fail("Expected a parser error");
-        }
-        catch (Exception e)
-        {
-            assertPureException(PureParserException.class, "Missing value for property 'type' on TypeView TestClass_1", 4, 14, e);
-        }
+        PureParserException e3 = Assert.assertThrows(PureParserException.class, () -> compileTestSource(
+                "###Diagram\n" +
+                        "Diagram test::pure::TestDiagram(width=10.0, height=10.0)\n" +
+                        "{\n" +
+                        "    TypeView TestClass_1(stereotypesVisible=true, attributesVisible=true,\n" +
+                        "                         attributeStereotypesVisible=true, attributeTypesVisible=true,\n" +
+                        "                         color=#FFFFCC, lineWidth=1.0,\n" +
+                        "                         position=(874.0, 199.46875), width=353.0, height=57.1875)\n" +
+                        "}\n"));
+        assertPureException(PureParserException.class, "Missing value for property 'type' on TypeView TestClass_1", 4, 14, e3);
 
-        try
-        {
-            compileTestSource("###Diagram\n" +
-                    "Diagram test::pure::TestDiagram(width=10.0, height=10.0)\n" +
-                    "{\n" +
-                    "    TypeView TestClass_1(type=test::pure::TestClass, stereotypesVisible=true, attributesVisible=true,\n" +
-                    "                         attributeStereotypesVisible=true, attributeTypesVisible=true,\n" +
-                    "                         color=#FFFFCC,\n" +
-                    "                         position=(874.0, 199.46875), width=353.0, height=57.1875)\n" +
-                    "}\n");
-            Assert.fail("Expected a parser error");
-        }
-        catch (Exception e)
-        {
-            assertPureException(PureParserException.class, "Missing value for property 'lineWidth' on TypeView TestClass_1", 4, 14, e);
-        }
+        PureParserException e4 = Assert.assertThrows(PureParserException.class, () -> compileTestSource(
+                "###Diagram\n" +
+                        "Diagram test::pure::TestDiagram(width=10.0, height=10.0)\n" +
+                        "{\n" +
+                        "    TypeView TestClass_1(type=test::pure::TestClass, stereotypesVisible=true, attributesVisible=true,\n" +
+                        "                         attributeStereotypesVisible=true, attributeTypesVisible=true,\n" +
+                        "                         color=#FFFFCC,\n" +
+                        "                         position=(874.0, 199.46875), width=353.0, height=57.1875)\n" +
+                        "}\n"));
+        assertPureException(PureParserException.class, "Missing value for property 'lineWidth' on TypeView TestClass_1", 4, 14, e4);
     }
 
     @Test
-    public void testDiagramWithInvalidAssociationView() throws NoSuchFieldException, IllegalAccessException
+    public void testDiagramWithInvalidAssociationView()
     {
-        Parser parser = getRuntimeDiagramParser();
         // Empty association view
-        try
-        {
-            compileTestSource("###Diagram\n" +
-                    "Diagram test::pure::TestDiagram(width=10.0, height=10.0)\n" +
-                    "{\n" +
-                    "    AssociationView TestAssociation_1()\n" +
-                    "}\n");
-            Assert.fail("Expected a parser error");
-        }
-        catch (Exception e)
-        {
-            assertPureException(PureParserException.class, "expected: one of {ASSOCIATION, STEREOTYPES_VISIBLE, NAME_VISIBLE, COLOR, LINE_WIDTH, LINE_STYLE, POINTS, LABEL, SOURCE, TARGET, SOURCE_PROP_POSITION, SOURCE_MULT_POSITION, TARGET_PROP_POSITION, TARGET_MULT_POSITION} found: ')'", 4, 39, e);
-        }
+        PureParserException e1 = Assert.assertThrows(PureParserException.class, () -> compileTestSource(
+                "###Diagram\n" +
+                        "Diagram test::pure::TestDiagram(width=10.0, height=10.0)\n" +
+                        "{\n" +
+                        "    AssociationView TestAssociation_1()\n" +
+                        "}\n"));
+        assertPureException(PureParserException.class, "expected: one of {ASSOCIATION, STEREOTYPES_VISIBLE, NAME_VISIBLE, COLOR, LINE_WIDTH, LINE_STYLE, POINTS, LABEL, SOURCE, TARGET, SOURCE_PROP_POSITION, SOURCE_MULT_POSITION, TARGET_PROP_POSITION, TARGET_MULT_POSITION} found: ')'", 4, 39, e1);
 
         // Association view with one bogus property
-        try
-        {
-            compileTestSource("###Diagram\n" +
-                    "Diagram test::pure::TestDiagram(width=10.0, height=10.0)\n" +
-                    "{\n" +
-                    "    AssociationView TestAssociation_1(junk=13.6)\n" +
-                    "}\n");
-            Assert.fail("Expected a parser error");
-        }
-        catch (Exception e)
-        {
-            assertPureException(PureParserException.class, "token recognition error at: '='", 4, 43, e);
-        }
+        PureParserException e2 = Assert.assertThrows(PureParserException.class, () -> compileTestSource(
+                "###Diagram\n" +
+                        "Diagram test::pure::TestDiagram(width=10.0, height=10.0)\n" +
+                        "{\n" +
+                        "    AssociationView TestAssociation_1(junk=13.6)\n" +
+                        "}\n"));
+        assertPureException(PureParserException.class, "token recognition error at: '='", 4, 43, e2);
 
         // Association view with all valid properties but one
-        try
-        {
-            compileTestSource("###Diagram\n" +
-                    "Diagram test::pure::TestDiagram(width=10.0, height=10.0)\n" +
-                    "{\n" +
-                    "    TypeView TestClass1_1(type=test::pure::TestClass1, stereotypesVisible=true, attributesVisible=true,\n" +
-                    "                          attributeStereotypesVisible=true, attributeTypesVisible=true,\n" +
-                    "                          color=#FFFFCC, lineWidth=1.0,\n" +
-                    "                          position=(874.0, 199.46875), width=353.0, height=57.1875)\n" +
-                    "    TypeView TestClass2_2(type=test::pure::TestClass2, stereotypesVisible=true, attributesVisible=true,\n" +
-                    "                          attributeStereotypesVisible=true, attributeTypesVisible=true,\n" +
-                    "                          color=#FFFFCC, lineWidth=1.0,\n" +
-                    "                          position=(75.0, 97.1875), width=113.0, height=57.1875)\n" +
-                    "    AssociationView TestAssociation_1(stereotypesVisible=true, nameVisible=false,\n" +
-                    "                                      color=#000000, lineWidth=1.0,\n" +
-                    "                                      lineStyle=SIMPLE, points=[(132.5, 77.0), (155.2, 77.0)],\n" +
-                    "                                      label='Employment',\n" +
-                    "                                      source=TestClass1_1,\n" +
-                    "                                      target=TestClass2_2,\n" +
-                    "                                      sourcePropertyPosition=(132.5, 76.2),\n" +
-                    "                                      sourceMultiplicityPosition=(132.5, 80.0),\n" +
-                    "                                      targetPropertyPosition=(155.2, 76.2),\n" +
-                    "                                      targetMultiplicityPosition=(155.2, 80.0))\n" +
-                    "}\n");
-            Assert.fail("Expected a parser error");
-        }
-        catch (Exception e)
-        {
-            assertPureException(PureParserException.class, "Missing value for property 'association' on AssociationView TestAssociation_1", 12, 21, e);
-        }
+        PureParserException e3 = Assert.assertThrows(PureParserException.class, () -> compileTestSource(
+                "###Diagram\n" +
+                        "Diagram test::pure::TestDiagram(width=10.0, height=10.0)\n" +
+                        "{\n" +
+                        "    TypeView TestClass1_1(type=test::pure::TestClass1, stereotypesVisible=true, attributesVisible=true,\n" +
+                        "                          attributeStereotypesVisible=true, attributeTypesVisible=true,\n" +
+                        "                          color=#FFFFCC, lineWidth=1.0,\n" +
+                        "                          position=(874.0, 199.46875), width=353.0, height=57.1875)\n" +
+                        "    TypeView TestClass2_2(type=test::pure::TestClass2, stereotypesVisible=true, attributesVisible=true,\n" +
+                        "                          attributeStereotypesVisible=true, attributeTypesVisible=true,\n" +
+                        "                          color=#FFFFCC, lineWidth=1.0,\n" +
+                        "                          position=(75.0, 97.1875), width=113.0, height=57.1875)\n" +
+                        "    AssociationView TestAssociation_1(stereotypesVisible=true, nameVisible=false,\n" +
+                        "                                      color=#000000, lineWidth=1.0,\n" +
+                        "                                      lineStyle=SIMPLE, points=[(132.5, 77.0), (155.2, 77.0)],\n" +
+                        "                                      label='Employment',\n" +
+                        "                                      source=TestClass1_1,\n" +
+                        "                                      target=TestClass2_2,\n" +
+                        "                                      sourcePropertyPosition=(132.5, 76.2),\n" +
+                        "                                      sourceMultiplicityPosition=(132.5, 80.0),\n" +
+                        "                                      targetPropertyPosition=(155.2, 76.2),\n" +
+                        "                                      targetMultiplicityPosition=(155.2, 80.0))\n" +
+                        "}\n"));
+        assertPureException(PureParserException.class, "Missing value for property 'association' on AssociationView TestAssociation_1", 12, 21, e3);
 
-        try
-        {
-            compileTestSource("###Diagram\n" +
-                    "Diagram test::pure::TestDiagram(width=10.0, height=10.0)\n" +
-                    "{\n" +
-                    "    TypeView TestClass1_1(type=test::pure::TestClass1, stereotypesVisible=true, attributesVisible=true,\n" +
-                    "                          attributeStereotypesVisible=true, attributeTypesVisible=true,\n" +
-                    "                          color=#FFFFCC, lineWidth=1.0,\n" +
-                    "                          position=(874.0, 199.46875), width=353.0, height=57.1875)\n" +
-                    "    TypeView TestClass2_2(type=test::pure::TestClass2, stereotypesVisible=true, attributesVisible=true,\n" +
-                    "                          attributeStereotypesVisible=true, attributeTypesVisible=true,\n" +
-                    "                          color=#FFFFCC, lineWidth=1.0,\n" +
-                    "                          position=(75.0, 97.1875), width=113.0, height=57.1875)\n" +
-                    "    AssociationView TestAssociation_1(association=test::pure::TestAssociation, stereotypesVisible=true, nameVisible=false,\n" +
-                    "                                      color=#000000,\n" +
-                    "                                      lineStyle=SIMPLE, points=[(132.5, 77.0), (155.2, 77.0)],\n" +
-                    "                                      label='Employment',\n" +
-                    "                                      source=TestClass1_1,\n" +
-                    "                                      target=TestClass2_2,\n" +
-                    "                                      sourcePropertyPosition=(132.5, 76.2),\n" +
-                    "                                      sourceMultiplicityPosition=(132.5, 80.0),\n" +
-                    "                                      targetPropertyPosition=(155.2, 76.2),\n" +
-                    "                                      targetMultiplicityPosition=(155.2, 80.0))\n" +
-                    "}\n");
-            Assert.fail("Expected a parser error");
-        }
-        catch (Exception e)
-        {
-            assertPureException(PureParserException.class, "Missing value for property 'lineWidth' on AssociationView TestAssociation_1", 12, 21, e);
-        }
+        PureParserException e4 = Assert.assertThrows(PureParserException.class, () -> compileTestSource(
+                "###Diagram\n" +
+                        "Diagram test::pure::TestDiagram(width=10.0, height=10.0)\n" +
+                        "{\n" +
+                        "    TypeView TestClass1_1(type=test::pure::TestClass1, stereotypesVisible=true, attributesVisible=true,\n" +
+                        "                          attributeStereotypesVisible=true, attributeTypesVisible=true,\n" +
+                        "                          color=#FFFFCC, lineWidth=1.0,\n" +
+                        "                          position=(874.0, 199.46875), width=353.0, height=57.1875)\n" +
+                        "    TypeView TestClass2_2(type=test::pure::TestClass2, stereotypesVisible=true, attributesVisible=true,\n" +
+                        "                          attributeStereotypesVisible=true, attributeTypesVisible=true,\n" +
+                        "                          color=#FFFFCC, lineWidth=1.0,\n" +
+                        "                          position=(75.0, 97.1875), width=113.0, height=57.1875)\n" +
+                        "    AssociationView TestAssociation_1(association=test::pure::TestAssociation, stereotypesVisible=true, nameVisible=false,\n" +
+                        "                                      color=#000000,\n" +
+                        "                                      lineStyle=SIMPLE, points=[(132.5, 77.0), (155.2, 77.0)],\n" +
+                        "                                      label='Employment',\n" +
+                        "                                      source=TestClass1_1,\n" +
+                        "                                      target=TestClass2_2,\n" +
+                        "                                      sourcePropertyPosition=(132.5, 76.2),\n" +
+                        "                                      sourceMultiplicityPosition=(132.5, 80.0),\n" +
+                        "                                      targetPropertyPosition=(155.2, 76.2),\n" +
+                        "                                      targetMultiplicityPosition=(155.2, 80.0))\n" +
+                        "}\n"));
+        assertPureException(PureParserException.class, "Missing value for property 'lineWidth' on AssociationView TestAssociation_1", 12, 21, e4);
     }
 
     @Test
-    public void testDiagramWithInvalidPropertyView() throws NoSuchFieldException, IllegalAccessException
+    public void testDiagramWithInvalidPropertyView()
     {
-        Parser parser = getRuntimeDiagramParser();
         // Create class with property
         compileTestSource("Class test::pure::TestClass1\n" +
                 "{\n" +
@@ -314,181 +206,132 @@ public class TestDiagramParsing extends AbstractPureTestWithCoreCompiled
                 "}\n");
 
         // Empty property view
-        try
-        {
-            compileTestSource("###Diagram\n" +
-                    "Diagram test::pure::TestDiagram(width=10.0, height=10.0)\n" +
-                    "{\n" +
-                    "    PropertyView TestClass1_testProperty_1()\n" +
-                    "}\n");
-            Assert.fail("Expected a parser error");
-        }
-        catch (Exception e)
-        {
-            assertPureException(PureParserException.class, "expected: one of {PROPERTY, STEREOTYPES_VISIBLE, NAME_VISIBLE, COLOR, LINE_WIDTH, LINE_STYLE, POINTS, LABEL, SOURCE, TARGET, PROP_POSITION, MULT_POSITION} found: ')'", 4, 44, e);
-        }
+        PureParserException e1 = Assert.assertThrows(PureParserException.class, () -> compileTestSource(
+                "###Diagram\n" +
+                        "Diagram test::pure::TestDiagram(width=10.0, height=10.0)\n" +
+                        "{\n" +
+                        "    PropertyView TestClass1_testProperty_1()\n" +
+                        "}\n"));
+        assertPureException(PureParserException.class, "expected: one of {PROPERTY, STEREOTYPES_VISIBLE, NAME_VISIBLE, COLOR, LINE_WIDTH, LINE_STYLE, POINTS, LABEL, SOURCE, TARGET, PROP_POSITION, MULT_POSITION} found: ')'", 4, 44, e1);
 
         // Property view with one bogus property
-        try
-        {
-            compileTestSource("###Diagram\n" +
-                    "Diagram test::pure::TestDiagram(width=10.0, height=10.0)\n" +
-                    "{\n" +
-                    "    PropertyView TestClass1_testProperty_1(junk=13.6)\n" +
-                    "}\n");
-            Assert.fail("Expected a parser error");
-        }
-        catch (Exception e)
-        {
-            assertPureException(PureParserException.class, "token recognition error at: '='", 4, 48, e);
-        }
+        PureParserException e2 = Assert.assertThrows(PureParserException.class, () -> compileTestSource(
+                "###Diagram\n" +
+                        "Diagram test::pure::TestDiagram(width=10.0, height=10.0)\n" +
+                        "{\n" +
+                        "    PropertyView TestClass1_testProperty_1(junk=13.6)\n" +
+                        "}\n"));
+        assertPureException(PureParserException.class, "token recognition error at: '='", 4, 48, e2);
 
         // Property view with all valid properties but one
-        try
-        {
-            compileTestSource("###Diagram\n" +
-                    "Diagram test::pure::TestDiagram(width=10.0, height=10.0)\n" +
-                    "{\n" +
-                    "    TypeView TestClass1_1(type=test::pure::TestClass1, stereotypesVisible=true, attributesVisible=true,\n" +
-                    "                          attributeStereotypesVisible=true, attributeTypesVisible=true,\n" +
-                    "                          color=#FFFFCC, lineWidth=1.0,\n" +
-                    "                          position=(874.0, 199.46875), width=353.0, height=57.1875)\n" +
-                    "    TypeView TestClass2_2(type=test::pure::TestClass2, stereotypesVisible=true, attributesVisible=true,\n" +
-                    "                          attributeStereotypesVisible=true, attributeTypesVisible=true,\n" +
-                    "                          color=#FFFFCC, lineWidth=1.0,\n" +
-                    "                          position=(75.0, 97.1875), width=113.0, height=57.1875)\n" +
-                    "    PropertyView TestClass1_testProperty_1(stereotypesVisible=true, nameVisible=false,\n" +
-                    "                                           color=#000000, lineWidth=1.0,\n" +
-                    "                                           lineStyle=SIMPLE, points=[(132.5, 77.0), (155.2, 77.0)],\n" +
-                    "                                           label='Employment',\n" +
-                    "                                           source=TestClass1_1,\n" +
-                    "                                           target=TestClass2_2,\n" +
-                    "                                           propertyPosition=(132.5, 76.2),\n" +
-                    "                                           multiplicityPosition=(132.5, 80.0))\n" +
-                    "}\n");
-            Assert.fail("Expected a parser error");
-        }
-        catch (Exception e)
-        {
-            assertPureException(PureParserException.class, "Missing value for property 'property' on PropertyView TestClass1_testProperty_1", 12, 18, e);
-        }
+        PureParserException e3 = Assert.assertThrows(PureParserException.class, () -> compileTestSource(
+                "###Diagram\n" +
+                        "Diagram test::pure::TestDiagram(width=10.0, height=10.0)\n" +
+                        "{\n" +
+                        "    TypeView TestClass1_1(type=test::pure::TestClass1, stereotypesVisible=true, attributesVisible=true,\n" +
+                        "                          attributeStereotypesVisible=true, attributeTypesVisible=true,\n" +
+                        "                          color=#FFFFCC, lineWidth=1.0,\n" +
+                        "                          position=(874.0, 199.46875), width=353.0, height=57.1875)\n" +
+                        "    TypeView TestClass2_2(type=test::pure::TestClass2, stereotypesVisible=true, attributesVisible=true,\n" +
+                        "                          attributeStereotypesVisible=true, attributeTypesVisible=true,\n" +
+                        "                          color=#FFFFCC, lineWidth=1.0,\n" +
+                        "                          position=(75.0, 97.1875), width=113.0, height=57.1875)\n" +
+                        "    PropertyView TestClass1_testProperty_1(stereotypesVisible=true, nameVisible=false,\n" +
+                        "                                           color=#000000, lineWidth=1.0,\n" +
+                        "                                           lineStyle=SIMPLE, points=[(132.5, 77.0), (155.2, 77.0)],\n" +
+                        "                                           label='Employment',\n" +
+                        "                                           source=TestClass1_1,\n" +
+                        "                                           target=TestClass2_2,\n" +
+                        "                                           propertyPosition=(132.5, 76.2),\n" +
+                        "                                           multiplicityPosition=(132.5, 80.0))\n" +
+                        "}\n"));
+        assertPureException(PureParserException.class, "Missing value for property 'property' on PropertyView TestClass1_testProperty_1", 12, 18, e3);
 
-        try
-        {
-            compileTestSource("###Diagram\n" +
-                    "Diagram test::pure::TestDiagram(width=10.0, height=10.0)\n" +
-                    "{\n" +
-                    "    TypeView TestClass1_1(type=test::pure::TestClass1, stereotypesVisible=true, attributesVisible=true,\n" +
-                    "                          attributeStereotypesVisible=true, attributeTypesVisible=true,\n" +
-                    "                          color=#FFFFCC, lineWidth=1.0,\n" +
-                    "                          position=(874.0, 199.46875), width=353.0, height=57.1875)\n" +
-                    "    TypeView TestClass2_2(type=test::pure::TestClass2, stereotypesVisible=true, attributesVisible=true,\n" +
-                    "                          attributeStereotypesVisible=true, attributeTypesVisible=true,\n" +
-                    "                          color=#FFFFCC, lineWidth=1.0,\n" +
-                    "                          position=(75.0, 97.1875), width=113.0, height=57.1875)\n" +
-                    "    PropertyView TestClass1_testProperty_1(property=test::pure::TestClass1.testProperty, stereotypesVisible=true, nameVisible=false,\n" +
-                    "                                           color=#000000,\n" +
-                    "                                           lineStyle=SIMPLE, points=[(132.5, 77.0), (155.2, 77.0)],\n" +
-                    "                                           label='Employment',\n" +
-                    "                                           source=TestClass1_1,\n" +
-                    "                                           target=TestClass2_2,\n" +
-                    "                                           propertyPosition=(132.5, 76.2),\n" +
-                    "                                           multiplicityPosition=(132.5, 80.0))\n" +
-                    "}\n");
-            Assert.fail("Expected a parser error");
-        }
-        catch (Exception e)
-        {
-            assertPureException(PureParserException.class, "Missing value for property 'lineWidth' on PropertyView TestClass1_testProperty_1", 12, 18, e);
-        }
+        PureParserException e4 = Assert.assertThrows(PureParserException.class, () -> compileTestSource(
+                "###Diagram\n" +
+                        "Diagram test::pure::TestDiagram(width=10.0, height=10.0)\n" +
+                        "{\n" +
+                        "    TypeView TestClass1_1(type=test::pure::TestClass1, stereotypesVisible=true, attributesVisible=true,\n" +
+                        "                          attributeStereotypesVisible=true, attributeTypesVisible=true,\n" +
+                        "                          color=#FFFFCC, lineWidth=1.0,\n" +
+                        "                          position=(874.0, 199.46875), width=353.0, height=57.1875)\n" +
+                        "    TypeView TestClass2_2(type=test::pure::TestClass2, stereotypesVisible=true, attributesVisible=true,\n" +
+                        "                          attributeStereotypesVisible=true, attributeTypesVisible=true,\n" +
+                        "                          color=#FFFFCC, lineWidth=1.0,\n" +
+                        "                          position=(75.0, 97.1875), width=113.0, height=57.1875)\n" +
+                        "    PropertyView TestClass1_testProperty_1(property=test::pure::TestClass1.testProperty, stereotypesVisible=true, nameVisible=false,\n" +
+                        "                                           color=#000000,\n" +
+                        "                                           lineStyle=SIMPLE, points=[(132.5, 77.0), (155.2, 77.0)],\n" +
+                        "                                           label='Employment',\n" +
+                        "                                           source=TestClass1_1,\n" +
+                        "                                           target=TestClass2_2,\n" +
+                        "                                           propertyPosition=(132.5, 76.2),\n" +
+                        "                                           multiplicityPosition=(132.5, 80.0))\n" +
+                        "}\n"));
+        assertPureException(PureParserException.class, "Missing value for property 'lineWidth' on PropertyView TestClass1_testProperty_1", 12, 18, e4);
     }
 
     @Test
-    public void testDiagramWithInvalidGeneralizationView() throws NoSuchFieldException, IllegalAccessException
+    public void testDiagramWithInvalidGeneralizationView()
     {
-        Parser parser = getRuntimeDiagramParser();
         // Empty generalization view
-        try
-        {
-            compileTestSource("###Diagram\n" +
-                    "Diagram test::pure::TestDiagram(width=10.0, height=10.0)\n" +
-                    "{\n" +
-                    "    GeneralizationView TestClass1_TestClass2_1()\n" +
-                    "}\n");
-            Assert.fail("Expected a parser error");
-        }
-        catch (Exception e)
-        {
-            assertPureException(PureParserException.class, "expected: one of {COLOR, LINE_WIDTH, LINE_STYLE, POINTS, LABEL, SOURCE, TARGET} found: ')'", 4, 48, e);
-        }
+        PureParserException e1 = Assert.assertThrows(PureParserException.class, () -> compileTestSource(
+                "###Diagram\n" +
+                        "Diagram test::pure::TestDiagram(width=10.0, height=10.0)\n" +
+                        "{\n" +
+                        "    GeneralizationView TestClass1_TestClass2_1()\n" +
+                        "}\n"));
+        assertPureException(PureParserException.class, "expected: one of {COLOR, LINE_WIDTH, LINE_STYLE, POINTS, LABEL, SOURCE, TARGET} found: ')'", 4, 48, e1);
 
         // Association view with one bogus property
-        try
-        {
-            compileTestSource("###Diagram\n" +
-                    "Diagram test::pure::TestDiagram(width=10.0, height=10.0)\n" +
-                    "{\n" +
-                    "    GeneralizationView TestClass1_TestClass2_1(junk=13.6)\n" +
-                    "}\n");
-            Assert.fail("Expected a parser error");
-        }
-        catch (Exception e)
-        {
-            assertPureException(PureParserException.class, "token recognition error at: '='", 4, 52, e);
-        }
+        PureParserException e2 = Assert.assertThrows(PureParserException.class, () -> compileTestSource(
+                "###Diagram\n" +
+                        "Diagram test::pure::TestDiagram(width=10.0, height=10.0)\n" +
+                        "{\n" +
+                        "    GeneralizationView TestClass1_TestClass2_1(junk=13.6)\n" +
+                        "}\n"));
+        assertPureException(PureParserException.class, "token recognition error at: '='", 4, 52, e2);
 
         // Association view with all valid properties but one
-        try
-        {
-            compileTestSource("###Diagram\n" +
-                    "Diagram test::pure::TestDiagram(width=10.0, height=10.0)\n" +
-                    "{\n" +
-                    "    TypeView TestClass1_1(type=test::pure::TestClass1, stereotypesVisible=true, attributesVisible=true,\n" +
-                    "                          attributeStereotypesVisible=true, attributeTypesVisible=true,\n" +
-                    "                          color=#FFFFCC, lineWidth=1.0,\n" +
-                    "                          position=(874.0, 199.46875), width=353.0, height=57.1875)\n" +
-                    "    TypeView TestClass2_2(type=test::pure::TestClass2, stereotypesVisible=true, attributesVisible=true,\n" +
-                    "                          attributeStereotypesVisible=true, attributeTypesVisible=true,\n" +
-                    "                          color=#FFFFCC, lineWidth=1.0,\n" +
-                    "                          position=(75.0, 97.1875), width=113.0, height=57.1875)\n" +
-                    "    GeneralizationView TestClass1_TestClass2_1(lineWidth=1.0,\n" +
-                    "                                               lineStyle=SIMPLE, points=[(132.5, 77.0), (155.2, 77.0)],\n" +
-                    "                                               label='',\n" +
-                    "                                               source=TestClass1_1,\n" +
-                    "                                               target=TestClass2_2)\n" +
-                    "}\n");
-            Assert.fail("Expected a parser error");
-        }
-        catch (Exception e)
-        {
-            assertPureException(PureParserException.class, "Missing value for property 'color' on GeneralizationView TestClass1_TestClass2_1", 12, 24, e);
-        }
+        PureParserException e3 = Assert.assertThrows(PureParserException.class, () -> compileTestSource(
+                "###Diagram\n" +
+                        "Diagram test::pure::TestDiagram(width=10.0, height=10.0)\n" +
+                        "{\n" +
+                        "    TypeView TestClass1_1(type=test::pure::TestClass1, stereotypesVisible=true, attributesVisible=true,\n" +
+                        "                          attributeStereotypesVisible=true, attributeTypesVisible=true,\n" +
+                        "                          color=#FFFFCC, lineWidth=1.0,\n" +
+                        "                          position=(874.0, 199.46875), width=353.0, height=57.1875)\n" +
+                        "    TypeView TestClass2_2(type=test::pure::TestClass2, stereotypesVisible=true, attributesVisible=true,\n" +
+                        "                          attributeStereotypesVisible=true, attributeTypesVisible=true,\n" +
+                        "                          color=#FFFFCC, lineWidth=1.0,\n" +
+                        "                          position=(75.0, 97.1875), width=113.0, height=57.1875)\n" +
+                        "    GeneralizationView TestClass1_TestClass2_1(lineWidth=1.0,\n" +
+                        "                                               lineStyle=SIMPLE, points=[(132.5, 77.0), (155.2, 77.0)],\n" +
+                        "                                               label='',\n" +
+                        "                                               source=TestClass1_1,\n" +
+                        "                                               target=TestClass2_2)\n" +
+                        "}\n"));
+        assertPureException(PureParserException.class, "Missing value for property 'color' on GeneralizationView TestClass1_TestClass2_1", 12, 24, e3);
 
-        try
-        {
-            compileTestSource("###Diagram\n" +
-                    "Diagram test::pure::TestDiagram(width=10.0, height=10.0)\n" +
-                    "{\n" +
-                    "    TypeView TestClass1_1(type=test::pure::TestClass1, stereotypesVisible=true, attributesVisible=true,\n" +
-                    "                          attributeStereotypesVisible=true, attributeTypesVisible=true,\n" +
-                    "                          color=#FFFFCC, lineWidth=1.0,\n" +
-                    "                          position=(874.0, 199.46875), width=353.0, height=57.1875)\n" +
-                    "    TypeView TestClass2_2(type=test::pure::TestClass2, stereotypesVisible=true, attributesVisible=true,\n" +
-                    "                          attributeStereotypesVisible=true, attributeTypesVisible=true,\n" +
-                    "                          color=#FFFFCC, lineWidth=1.0,\n" +
-                    "                          position=(75.0, 97.1875), width=113.0, height=57.1875)\n" +
-                    "    GeneralizationView TestClass1_TestClass2_1(color=#000000,\n" +
-                    "                                               lineStyle=SIMPLE, points=[(132.5, 77.0), (155.2, 77.0)],\n" +
-                    "                                               label='',\n" +
-                    "                                               source=TestClass1_1,\n" +
-                    "                                               target=TestClass2_2)\n" +
-                    "}\n");
-            Assert.fail("Expected a parser error");
-        }
-        catch (Exception e)
-        {
-            assertPureException(PureParserException.class, "Missing value for property 'lineWidth' on GeneralizationView TestClass1_TestClass2_1", 12, 24, e);
-        }
+        PureParserException e4 = Assert.assertThrows(PureParserException.class, () -> compileTestSource(
+                "###Diagram\n" +
+                        "Diagram test::pure::TestDiagram(width=10.0, height=10.0)\n" +
+                        "{\n" +
+                        "    TypeView TestClass1_1(type=test::pure::TestClass1, stereotypesVisible=true, attributesVisible=true,\n" +
+                        "                          attributeStereotypesVisible=true, attributeTypesVisible=true,\n" +
+                        "                          color=#FFFFCC, lineWidth=1.0,\n" +
+                        "                          position=(874.0, 199.46875), width=353.0, height=57.1875)\n" +
+                        "    TypeView TestClass2_2(type=test::pure::TestClass2, stereotypesVisible=true, attributesVisible=true,\n" +
+                        "                          attributeStereotypesVisible=true, attributeTypesVisible=true,\n" +
+                        "                          color=#FFFFCC, lineWidth=1.0,\n" +
+                        "                          position=(75.0, 97.1875), width=113.0, height=57.1875)\n" +
+                        "    GeneralizationView TestClass1_TestClass2_1(color=#000000,\n" +
+                        "                                               lineStyle=SIMPLE, points=[(132.5, 77.0), (155.2, 77.0)],\n" +
+                        "                                               label='',\n" +
+                        "                                               source=TestClass1_1,\n" +
+                        "                                               target=TestClass2_2)\n" +
+                        "}\n"));
+        assertPureException(PureParserException.class, "Missing value for property 'lineWidth' on GeneralizationView TestClass1_TestClass2_1", 12, 24, e4);
     }
 
     @Test
@@ -504,12 +347,12 @@ public class TestDiagramParsing extends AbstractPureTestWithCoreCompiled
 
         CoreInstance width = Instance.getValueForMetaPropertyToOneResolved(geometry, "width", processorSupport);
         Assert.assertNotNull(width);
-        Assert.assertTrue(Instance.instanceOf(width, "Float", processorSupport));
+        Assert.assertTrue(Instance.instanceOf(width, M3Paths.Float, processorSupport));
         Assert.assertEquals("0.0", width.getName());
 
         CoreInstance height = Instance.getValueForMetaPropertyToOneResolved(geometry, "height", processorSupport);
         Assert.assertNotNull(height);
-        Assert.assertTrue(Instance.instanceOf(height, "Float", processorSupport));
+        Assert.assertTrue(Instance.instanceOf(height, M3Paths.Float, processorSupport));
         Assert.assertEquals("0.0", height.getName());
     }
 
@@ -606,7 +449,7 @@ public class TestDiagramParsing extends AbstractPureTestWithCoreCompiled
     @Test
     public void testDiagramModelDiagram()
     {
-        final String source = "###Diagram\n" +
+        String source = "###Diagram\n" +
                 "Diagram meta::pure::diagram::DiagramDiagram(width=924.0, height=798.0)\n" +
                 "{\n" +
                 "    TypeView AbstractPathView(type=meta::pure::diagram::AbstractPathView,\n" +
@@ -871,20 +714,12 @@ public class TestDiagramParsing extends AbstractPureTestWithCoreCompiled
                 "                                            source=TypeView,\n" +
                 "                                            target=DiagramNode)\n" +
                 "}";
-        compileTestSource("testDiagram.pure",
-                source);
+        compileTestSource("testDiagram.pure", source);
 
-        Class typeViewClass = (Class) runtime.getCoreInstance("meta::pure::diagram::TypeView");
-        RichIterable<? extends ReferenceUsage> typeViewReferenceUsages = typeViewClass._referenceUsages().select(new Predicate<ReferenceUsage>()
-        {
-            @Override
-            public boolean accept(ReferenceUsage usage)
-            {
-                return usage._owner() instanceof TypeView;
-            }
-        });
+        Class<?> typeViewClass = (Class<?>) runtime.getCoreInstance("meta::pure::diagram::TypeView");
+        RichIterable<? extends ReferenceUsage> typeViewReferenceUsages = typeViewClass._referenceUsages().select(usage -> usage._owner() instanceof TypeView);
 
-        String[] lines = source.split("\n");
+        String[] lines = source.split("\\R");
         for (ReferenceUsage referenceUsage : typeViewReferenceUsages)
         {
             SourceInformation sourceInformation = referenceUsage.getSourceInformation();
@@ -895,7 +730,7 @@ public class TestDiagramParsing extends AbstractPureTestWithCoreCompiled
     @Test
     public void testDiagramModelDiagramWithAssociationView()
     {
-        final String source = "###Diagram\n" +
+        String source = "###Diagram\n" +
                 "Diagram meta::pure::diagram::DiagramDiagram(width=924.0, height=798.0)\n" +
                 "{\n" +
                 "    TypeView AssociationView(type=meta::pure::diagram::AssociationView,\n" +
@@ -958,33 +793,17 @@ public class TestDiagramParsing extends AbstractPureTestWithCoreCompiled
                 "                            targetMultiplicityPosition=(450.33789, 291.70564))\n" +
                 "   } ";
         compileTestSource("testDiagram.pure", source);
-        String[] lines = source.split("\n");
+        String[] lines = source.split("\\R");
 
         Association associationView = (Association) runtime.getCoreInstance("meta::pure::diagram::DiagramAssociationViews");
-        RichIterable<? extends ReferenceUsage> associationViewReferenceUsages = associationView._referenceUsages().select(new Predicate<ReferenceUsage>()
-        {
-            @Override
-            public boolean accept(ReferenceUsage usage)
-            {
-                return usage._owner() instanceof AssociationView;
-            }
-        });
-        for (ReferenceUsage referenceUsage : associationViewReferenceUsages)
+        for (ReferenceUsage referenceUsage : associationView._referenceUsages().select(usage -> usage._owner() instanceof AssociationView))
         {
             SourceInformation sourceInformation = referenceUsage.getSourceInformation();
             Assert.assertEquals("DiagramAssociationViews", lines[sourceInformation.getLine() - 1].substring(sourceInformation.getColumn() - 1, sourceInformation.getColumn() + "DiagramAssociationViews".length() - 1));
         }
 
         associationView = (Association) runtime.getCoreInstance("meta::pure::diagram::DiagramTypeViews");
-        associationViewReferenceUsages = associationView._referenceUsages().select(new Predicate<ReferenceUsage>()
-        {
-            @Override
-            public boolean accept(ReferenceUsage usage)
-            {
-                return usage._owner() instanceof AssociationView;
-            }
-        });
-        for (ReferenceUsage referenceUsage : associationViewReferenceUsages)
+        for (ReferenceUsage referenceUsage : associationView._referenceUsages().select(usage -> usage._owner() instanceof AssociationView))
         {
             SourceInformation sourceInformation = referenceUsage.getSourceInformation();
             Assert.assertEquals("DiagramTypeViews", lines[sourceInformation.getLine() - 1].substring(sourceInformation.getColumn() - 1, sourceInformation.getColumn() + "DiagramTypeViews".length() - 1));
